@@ -1,16 +1,16 @@
 {
   "private": true,
-  "version": "3.0.0-alpha.14.3",
+  "version": "3.0.0-alpha.14.4.0",
   "dependencies": {},
   "devDependencies": {
     "assert": "~1.3.0",
     "axios": "^0.18.0",
     "babel-eslint": "^6.1.2",
     "chalk": "^2.4.1",
-    "eslint": "^3.12.2",
+    "eslint": "^4.19.1",
     "eslint-plugin-babel": "^4.0.0",
-    "eslint-plugin-react": "^6.8.0",
-    "eslint-plugin-redux-saga": "^0.3.0",
+    "eslint-plugin-react": "^7.7.0",
+    "eslint-plugin-redux-saga": "^0.8.0",
     "istanbul": "~0.4.2",
     "jest": "^22.4.3",
     "jest-cli": "^22.4.4",
