@@ -10,7 +10,9 @@
       @blur="focused = false"
       @keyup.enter="go(focusIndex)"
       @keyup.up="onUp"
-      @keyup.down="onDown">
+      @keyup.down="onDown"
+      placeholder="Search"
+    >
     <ul class="suggestions"
       v-if="showSuggestions"
       :class="{ 'align-right': alignRight }"
@@ -63,8 +65,10 @@ export default {
       for (let i = 0; i < pages.length; i++) {
         if (res.length >= max) break
         const p = pages[i]
-        // filter out results that do not match current ersion context
-        if (~p.path.slice(1).indexOf(searchContext)) continue
+        // filter out results that do not match current version context
+        if ( ! p.path.startsWith(searchContext, 1)) {
+          continue;
+        }
         // filter out results that do not match current locale
         if (this.getPageLocalePath(p) !== localePath) continue
         if (matches(p)) {
