@@ -14,14 +14,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Switch, Route } from 'react-router-dom';
-
 import AdminPage from 'containers/AdminPage';
 import NotFoundPage from 'containers/NotFoundPage';
-
 import NotificationProvider from 'containers/NotificationProvider';
-
+import AppLoader from 'containers/AppLoader';
+import LoadingIndicatorPage from 'components/LoadingIndicatorPage';
 import '../../styles/main.scss';
-
 import styles from './styles.scss';
 
 export class App extends React.Component { // eslint-disable-line react/prefer-stateless-function
@@ -29,12 +27,22 @@ export class App extends React.Component { // eslint-disable-line react/prefer-s
     return (
       <div>
         <NotificationProvider />
-        <div className={styles.container}>
-          <Switch>
-            <Route path="/" component={AdminPage} />
-            <Route path="" component={NotFoundPage} />
-          </Switch>
-        </div>
+        <AppLoader>
+          {({ shouldLoad }) => {
+            if (shouldLoad) {
+              return <LoadingIndicatorPage />;
+            }
+
+            return (
+              <div className={styles.container}>
+                <Switch>
+                  <Route path="/" component={AdminPage} />
+                  <Route path="" component={NotFoundPage} />
+                </Switch>
+              </div>
+            );
+          }}
+        </AppLoader>
       </div>
     );
   }
