@@ -10,6 +10,7 @@ const _ = require('lodash');
 const pluralize = require('pluralize');
 const policyUtils = require('strapi-utils').policy;
 const Query = require('./Query.js');
+/* eslint-disable no-unused-vars */
 
 module.exports = {
   /**
@@ -64,10 +65,7 @@ module.exports = {
         const [name, action] = handler.split('.');
 
         const controller = plugin
-          ? _.get(
-              strapi.plugins,
-              `${plugin}.controllers.${_.toLower(name)}.${action}`,
-            )
+          ? _.get(strapi.plugins, `${plugin}.controllers.${_.toLower(name)}.${action}`)
           : _.get(strapi.controllers, `${_.toLower(name)}.${action}`);
 
         if (!controller) {
@@ -148,10 +146,7 @@ module.exports = {
       const [name, action] = resolverOf.split('.');
 
       const controller = plugin
-        ? _.get(
-            strapi.plugins,
-            `${plugin}.controllers.${_.toLower(name)}.${action}`,
-          )
+        ? _.get(strapi.plugins, `${plugin}.controllers.${_.toLower(name)}.${action}`)
         : _.get(strapi.controllers, `${_.toLower(name)}.${action}`);
 
       if (!controller) {
