@@ -65,11 +65,11 @@ module.exports = {
         });
 
         Object.assign(acc.resolver[globalId], {
-          createdAt: (obj, options, context) => {
+          createdAt: (obj) => {
             // eslint-disable-line no-unused-vars
             return obj.createdAt || obj.created_at;
           },
-          updatedAt: (obj, options, context) => {
+          updatedAt: (obj) => {
             // eslint-disable-line no-unused-vars
             return obj.updatedAt || obj.updated_at;
           },
@@ -154,7 +154,7 @@ module.exports = {
       Object.keys(queries).forEach(type => {
         // The query cannot be built.
         if (_.isError(queries[type])) {
-          console.error(queries[type]);
+          strapi.log.error(queries[type]);
           strapi.stop();
         }
 
@@ -303,7 +303,7 @@ module.exports = {
           case 'manyMorphToMany':
           case 'manyToManyMorph':
             return _.merge(acc.resolver[globalId], {
-              [association.alias]: async (obj, options, context) => {
+              [association.alias]: async (obj) => {
                 // eslint-disable-line no-unused-vars
                 const [withRelated, withoutRelated] = await Promise.all([
                   resolvers.fetch(
@@ -362,7 +362,7 @@ module.exports = {
         }
 
         _.merge(acc.resolver[globalId], {
-          [association.alias]: async (obj, options, context) => {
+          [association.alias]: async (obj, options) => {
             // eslint-disable-line no-unused-vars
             // Construct parameters object to retrieve the correct related entries.
             const params = {
@@ -398,7 +398,7 @@ module.exports = {
               queryOpts.skip = convertedParams.start;
 
               switch (association.nature) {
-                case 'manyToMany': {
+                case 'manyToMany':
                   if (association.dominant) {
                     const arrayOfIds = (obj[association.alias] || []).map(
                       related => {
@@ -413,10 +413,10 @@ module.exports = {
                       [ref.primaryKey]: arrayOfIds,
                       ...where.where,
                     }).where;
+                    break;
+                  // falls through
                   }
                   break;
-                  // falls through
-                }
                 default:
                   // Where.
                   queryOpts.query = strapi.utils.models.convertParams(name, {
