@@ -63,7 +63,7 @@ module.exports = function (plugin, cliArguments) {
 
     if (!isStrapiInstalledWithNPM) {
       // Create the directory yarn doesn't do it it
-      shell.exec('mkdir', [pluginPath]);
+      shell.mkdir('-p', [pluginPath]);
       // Add a package.json so it installs the dependencies
       shell.touch(`${pluginPath}/package.json`);
       fs.writeFileSync(`${pluginPath}/package.json`, JSON.stringify({}), 'utf8');
